+++ configuration.properties
@@ -1,17 +1,18 @@
-browser = chrome
-url = https://qa.centrilli.com
+browser=chrome
+url=https://qa.centrilli.com
 
-POS_Manager.username =posmanager5@info.com
-POS_Manager.password =posmanager
+POS_Manager_username=posmanager5@info.com
+POS_Manager_password=posmanager
 
-CRM_Manager.username =eventscrmmanager10@info.com
-CRM_Manager.password =eventscrmmanager
+CRM_Manager_username=eventscrmmanager10@info.com
+CRM_Manager_password=eventscrmmanager
 
-Sales_Manager.password = salesmanager
+Sales_Manager_username=salesmanager6@info.com
+Sales_Manager_password=salesmanager
 
 
-Inventory_Manager.username =imm10@info.com
-Inventory_Manager.password =inventorymanager
+Inventory_Manager_username=imm10@info.com
+Inventory_Manager_password=inventorymanager
 
-Expenses_Manager.username =expensesmanager10@info.com
-Expenses_Manager.password =expensesmanager
+Expenses_Manager_username=expensesmanager10@info.com
+Expenses_Manager_password=expensesmanager
