+++ src/test/java/com/ERP/runners/CukesRunner.java
@@ -15,7 +15,7 @@ import org.junit.runner.RunWith;
         features = "src/test/resources/features",
         glue = "com/ERP/step_definitions",
         dryRun = false,
-        tags = "@first",
+        tags = "@wipH" ,
         publish = true //generating a report with public link
 )
